refactor(site): move `Add users` to header
